+++ HISTORY.rst
@@ -6,6 +6,11 @@ Unreleased Changes
 * Implement the proposed new function ``pygeoprocessing.raster_reduce``, a
   wrapper around ``pygeoprocessing.iterblocks``
    (https://github.com/natcap/pygeoprocessing/issues/285)
+* Added a D8 stream extraction function at
+  ``pygeoprocessing.routing.extract_streams_d8`` which takes a D8 flow
+  accumulation raster and a flow accumulation threshold, setting all pixels
+  with accumulation above that threshold to 1 and all other valid pixels to 0.
+  https://github.com/natcap/pygeoprocessing/issues/272
 * Adding a new function, ``pygeoprocessing.create_raster_from_bounding_box``,
   that enables the creation of a new raster from a bounding box.
   https://github.com/natcap/pygeoprocessing/issues/276